enhance: show preview of broken links for note deletion

+++ packages/plugin-core/src/commands/DeleteNodeCommand.ts
@@ -1,4 +1,5 @@
 import {
+  DLink,
   DVault,
   EngineDeletePayload,
   NoteProps,
@@ -6,8 +7,9 @@ import {
   SchemaUtils,
   VaultUtils,
 } from "@dendronhq/common-all";
+import _ from "lodash";
 import path from "path";
-import { TextEditor, window } from "vscode";
+import { TextEditor, ViewColumn, window } from "vscode";
 import { DendronClientUtilsV2 } from "../clientUtils";
 import { PickerUtilsV2 } from "../components/lookup/utils";
 import { DENDRON_COMMANDS } from "../constants";
@@ -15,9 +17,11 @@ import { Logger } from "../logger";
 import { VSCodeUtils } from "../vsCodeUtils";
 import { getDWorkspace, getEngine, getExtension } from "../workspace";
 import { BasicCommand } from "./base";
+import _md from "markdown-it";
 
 type CommandOpts = {
   _fsPath?: string;
+  noConfirm?: boolean;
 };
 
 type CommandOutput = EngineDeletePayload | void;
@@ -32,6 +36,8 @@ function formatDeletedMsg({
   return `${path.basename(fsPath)} (${VaultUtils.getName(vault)}) deleted`;
 }
 
+const md = _md();
+
 export class DeleteNodeCommand extends BasicCommand<
   CommandOpts,
   CommandOutput
@@ -39,6 +45,50 @@ export class DeleteNodeCommand extends BasicCommand<
   key = DENDRON_COMMANDS.DELETE_NODE.key;
   async gatherInputs(): Promise<any> {
     return {};
+  }
+
+  async showNoteDeletePreview(note: NoteProps, backlinks: DLink[]) {
+    let content = [
+      "# Delete Node Preview",
+      "```",
+      `node type: note`,
+      "",
+      `# of backlinks to this note: ${backlinks.length}`,
+      "```",
+      "## Broken links after deletion",
+      `These links will be broken after deleting _*${note.fname}*_`,
+      "",
+      `Make sure to convert the broken links listed below accordingly.`,
+      "",
+    ];
+
+    _.forEach(_.sortBy(backlinks, ["from.vaultName"]), (backlink) => {
+      const entry = [
+        `- in **${backlink.from.vaultName}/${backlink.from.fname}**`,
+        `  - line *${backlink.position?.start.line}* column *${backlink.position?.start.column}*`,
+        `  - alias: \`${backlink.alias ? backlink.alias : "None"}\``,
+      ].join("\n");
+      content = content.concat(entry);
+    });
+
+    const panel = window.createWebviewPanel(
+      "deleteNodeNoteDeletePreview",
+      "Note Delete Preview",
+      ViewColumn.One,
+      {}
+    );
+    panel.webview.html = md.render(content.join("\n"));
+  }
+
+  async promptConfirmation(title: string, noConfirm?: boolean) {
+    if (noConfirm) return true;
+    const options = ["Proceed", "Cancel"];
+    const resp = await VSCodeUtils.showQuickPick(options, {
+      title,
+      placeHolder: "Proceed",
+      ignoreFocusOut: true,
+    });
+    return resp === "Proceed";
   }
 
   async execute(opts?: CommandOpts): Promise<CommandOutput> {
@@ -54,13 +104,33 @@ export class DeleteNodeCommand extends BasicCommand<
       const fname = path.basename(fsPath, trimEnd);
       const client = getExtension().getEngine();
       if (mode === "note") {
-        const vault = PickerUtilsV2.getOrPromptVaultForOpenEditor();
+        const vault = PickerUtilsV2.getVaultForOpenEditor();
         const note = NoteUtils.getNoteByFnameV5({
           fname,
           vault,
           notes: getEngine().notes,
           wsRoot: getDWorkspace().wsRoot,
         }) as NoteProps;
+
+        const backlinks = note.links.filter((link) => link.type === "backlink");
+        let title;
+        if (backlinks.length === 0) {
+          // no need to show preview a simple
+          title = `Delete note ${note.fname}?`;
+        } else {
+          await this.showNoteDeletePreview(note, backlinks);
+          title = `${note.fname} has backlinks. Delete note?`;
+        }
+
+        const shouldProceed = await this.promptConfirmation(
+          title,
+          opts?.noConfirm
+        );
+        if (!shouldProceed) {
+          window.showInformationMessage("Cancelled");
+          return;
+        }
+
         const out = (await client.deleteNote(note.id)) as EngineDeletePayload;
         if (out.error) {
           Logger.error({ ctx, msg: "error deleting node", error: out.error });
